# specs/sep10/openapi.yaml
openapi: 3.1.0
info:
  title: SEP-10 Stellar Web Authentication
  version: 1.0.0
  description: API contract for SEP-10 challenge and token flow.
servers:
  - url: https://example.com
paths:
  /auth:
    get:
      operationId: getChallenge
      summary: Request SEP-10 challenge
      parameters:
        - in: query
          name: account
          required: true
          schema:
            type: string
            pattern: '^G[A-Z2-7]{55}$'
        - in: query
          name: home_domain
          required: false
          schema:
            type: string
        - in: query
          name: client_domain
          required: false
          schema:
            type: string
        - in: query
          name: memo
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Challenge transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      operationId: postChallenge
      summary: Verify signed challenge and get token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeVerifyRequest'
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  schemas:
    ChallengeResponse:
      type: object
      required: [transaction, network_passphrase]
      properties:
        transaction:
          type: string
        network_passphrase:
          type: string
    ChallengeVerifyRequest:
      type: object
      required: [transaction]
      properties:
        transaction:
          type: string
    TokenResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'